refactor(openworkflow): reorganize workflow structure and update imports
- Updated the directory structure for workflows, moving files into a dedicated 'workflows' folder.
- Adjusted import paths in various files to reflect the new structure.
- Enhanced the ignore patterns in the OpenWorkflow configuration to exclude test files.
- Introduced new workflows for Confluence sync operations, including content and space synchronization.
- Added repository ingestion workflow to streamline the process of syncing GitHub repositories.

This refactor aims to improve organization and maintainability of the workflow files.
