+++ apps/backend/src/openworkflow/workflows/confluence-sync-space.ts
@@ -1,10 +1,10 @@
 import { defineWorkflow } from "openworkflow"
 import { z } from "zod"
-import { parseEnv } from "../config/env.js"
-import { withOrgDbContext } from "../db/client.js"
-import { getForgeInstallationByConnectionId } from "../models/atlassian-connector.js"
-import { getConfluenceSyncTargetByConnectionId } from "../models/confluence-sync-target.js"
-import { syncConfluenceContent } from "../services/confluence/sync.js"
+import { parseEnv } from "../../config/env.js"
+import { withOrgDbContext } from "../../db/client.js"
+import { getForgeInstallationByConnectionId } from "../../models/atlassian-connector.js"
+import { getConfluenceSyncTargetByConnectionId } from "../../models/confluence-sync-target.js"
+import { syncConfluenceContent } from "../../services/confluence/sync.js"
 
 const confluenceSyncSpaceInputSchema = z.object({
   orgId: z.string().min(1),
